+++ R/deploy_flask-r6.R
@@ -1,0 +1,22 @@
+#' R6 classes
+#'
+#' Check classes
+#' @export
+check_deploy_python_flask = R6::R6Class(
+  "check_deploy_python_flask",
+  inherit = base_check,
+  public = list(
+    #' @description Checks deployment of Quarto document with HTML output
+    check = function() {
+      python_dir = dir = system.file("extdata", private$group, private$short,
+                               package = "jrHealthCheckConnect", mustWork = TRUE)
+      private$checker(deploy_flask(dir))
+      return(invisible(NULL))
+    }
+  ),
+  private = list(
+    context = "Deploy flask app",
+    short = "flask",
+    group = "deploy_python"
+  )
+)